+++ example/lib/ultimate/binder.dart
@@ -1,6 +1,5 @@
-/// * Description: 
+/// * Description:
 /// * Author:      郭文龙
 /// * Date:        2022/7/7 0:42
 /// * Email:       guowenlong20000@sina.com
-class binder {
-}
+class binder {}