+++ example/lib/ultimate/ultimate_page.dart
@@ -2,5 +2,4 @@
 /// * Author:      郭文龙
 /// * Date:        2022/7/1 18:24
 /// * Email:       guowenlong20000@sina.com
-class UltimatePage {
-}
+class UltimatePage {}