+++ CHANGELOG.md
@@ -1,2 +1,6 @@
-## [0.9.0] July 13, 2022
-- Complete `multitype` function:
+## [0.9.2] July 12, 2022
+- Format dart files.
+## [0.9.1] July 12, 2022
+- Make pub points up.
+## [0.9.0] July 10, 2022
+- Complete `multitype` function.